flappy/requirements.txt
mediapipe
pygame
pyglet
screeninfo
opencv-python
openpyxl
pandas
pillow
playsound